Demonstrate the 'Argument "attr value" isn't numeric' bug
When:
  - your role has an attribute which is lvalue
  - and has a string default
    - provided either as a scalar, or a subref returning one

using it in sprintf triggers the error "Argument
"attr value" isn't numeric in sprintf" - which is
nonsensical because the pattern for this argument is %s,
so we don't want it to be numeric.

"attr value" is the actual value of the attribute, so
if you set the value when constructing the object, it is
the value you provided, not the default value. If you
don't set the value, then it is the default provided.

Values which can be numified are unaffected.